requirements.txt
asgiref~=3.3.1
certifi~=2019.3.9
cffi==1.14.6
charset-normalizer==2.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django~=3.1.7
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna~=2.8
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib~=3.0.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz~=2018.9
requests~=2.21.0
requests-oauthlib==1.3.0
six~=1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse~=0.3.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
pip~=21.1.3
lxml~=4.6.2
Faker~=8.10.1
python-dateutil~=2.8.1
psycopg2~=2.8.4
numpy~=1.18.5
chardet~=3.0.4
setuptools~=40.8.0
Pygments~=2.3.1